feat: perform retry on error
